+++ src/generate/gen_flexgrid_sizer.cpp
@@ -1,7 +1,7 @@
 /////////////////////////////////////////////////////////////////////////////
 // Purpose:   wxFlexGridSizer generator
 // Author:    Ralph Walden
-// Copyright: Copyright (c) 2020-2023 KeyWorks Software (Ralph Walden)
+// Copyright: Copyright (c) 2020-2025 KeyWorks Software (Ralph Walden)
 // License:   Apache License -- see ../../LICENSE
 /////////////////////////////////////////////////////////////////////////////
 
@@ -71,8 +71,10 @@ bool FlexGridSizerGenerator::ConstructionCode(Code& code)
 
     // If rows is empty, only columns are supplied and wxFlexGridSizer will deduece the
     // number of rows to use
-    if (node->as_int(prop_rows) != 0)
+    if (node->as_int(prop_rows) != 0 || code.is_perl())
     {
+        // REVIEW: [Randalphwa - 04-22-2025] Currently, wxPerl requires all 4 parameters, unlike the
+        // other languages.
         code.as_string(prop_rows).Comma();
     }
     code.as_string(prop_cols).Comma().as_string(prop_vgap).Comma().as_string(prop_hgap).EndFunction();