+++ src/generate/gen_grid_sizer.cpp
@@ -1,7 +1,7 @@
 /////////////////////////////////////////////////////////////////////////////
 // Purpose:   wxGridSizer generator
 // Author:    Ralph Walden
-// Copyright: Copyright (c) 2020-2023 KeyWorks Software (Ralph Walden)
+// Copyright: Copyright (c) 2020-2025 KeyWorks Software (Ralph Walden)
 // License:   Apache License -- see ../../LICENSE
 /////////////////////////////////////////////////////////////////////////////
 
@@ -44,8 +44,10 @@ void GridSizerGenerator::AfterCreation(wxObject* wxobject, wxWindow* /*wxparent*
 bool GridSizerGenerator::ConstructionCode(Code& code)
 {
     code.AddAuto().NodeName().CreateClass();
-    if (code.node()->as_int(prop_rows) != 0)
+    if (code.node()->as_int(prop_rows) != 0 || code.is_perl())
     {
+        // REVIEW: [Randalphwa - 04-22-2025] Currently, wxPerl requires all 4 parameters, unlike the
+        // other languages.
         code.as_string(prop_rows).Comma();
     }
     code.as_string(prop_cols).Comma().as_string(prop_vgap).Comma().as_string(prop_hgap).EndFunction();