+++ src/generate/gen_std_dlgbtn_sizer.h
@@ -22,8 +22,4 @@ public:
 
     int GenXrcObject(Node*, pugi::xml_node& /* object */, size_t /* xrc_flags */) override;
     void RequiredHandlers(Node*, std::set<std::string>& /* handlers */) override;
-
-protected:
-    void GenPythonConstruction(Code& code);
-    void GenRubyConstruction(Code& code);
 };